+++ lib/openapi_first/test/coverage.rb
@@ -15,9 +15,10 @@ module OpenapiFirst
       @current_run = {}
 
       class << self
-        attr_reader :current_run
-
-        def install = Test.install
+        # @visibility private
+        def install
+          raise NoMethodError, 'Coverage.install was removed. Please use Test.setup instead'
+        end
 
         def start(skip_response: nil, skip_route: nil)
           @current_run = Test.definitions.values.to_h do |oad|
@@ -26,7 +27,10 @@ module OpenapiFirst
           end
         end
 
-        def uninstall = Test.uninstall
+        # @visibility private
+        def uninstall
+          raise NoMethodError, 'Coverage.uninstall was removed. Please use Test.uninstall instead'
+        end
 
         # Clear current coverage run
         def reset
@@ -45,12 +49,14 @@ module OpenapiFirst
           Result.new(plans:, coverage:)
         end
 
+        private
+
+        attr_reader :current_run
+
         # Returns all plans (Plan) that were registered for this run
         def plans
           current_run.values
         end
-
-        private
 
         def coverage
           return 0 if plans.empty?